rename combinatorial file to conform to object name
