+++ src/pynyu/discrete/common/combinatorial_number.py
